refactor: create separate module for each function
